Move ssladapter to transport module
