+++ docker/tls.py
@@ -1,7 +1,8 @@
 import os
 import ssl
 
-from . import errors, ssladapter
+from . import errors
+from .transport import SSLAdapter
 
 
 class TLSConfig(object):
@@ -84,7 +85,7 @@ class TLSConfig(object):
         if self.cert:
             client.cert = self.cert
 
-        client.mount('https://', ssladapter.SSLAdapter(
+        client.mount('https://', SSLAdapter(
             ssl_version=self.ssl_version,
             assert_hostname=self.assert_hostname,
             assert_fingerprint=self.assert_fingerprint,
