+++ docker/transport/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 from .unixconn import UnixAdapter
+from .ssladapter import SSLAdapter
 try:
     from .npipeconn import NpipeAdapter
     from .npipesocket import NpipeSocket
