+++ docker/transport/ssladapter.py
